Rename NofiticationDelivery to DeliveryPretreament.
