+++ Source/Microsoft.Teams.Apps.CompanyCommunicator.Send.Func/CompanyCommunicatorPretreatFunction.cs
@@ -18,24 +18,24 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Send.Func
     public class CompanyCommunicatorPretreatFunction
     {
         private readonly NotificationDataRepository notificationDataRepository;
-        private readonly NotificationDelivery.NotificationDelivery notificationDelivery;
+        private readonly DeliveryPretreatment.DeliveryPretreatment deliveryPretreatment;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="CompanyCommunicatorPretreatFunction"/> class.
         /// </summary>
         /// <param name="notificationDataRepository">Notification data repository service that deals with the table storage in azure.</param>
-        /// <param name="notificationDelivery">Notification delivery service instance.</param>
+        /// <param name="deliveryPretreatment">Notification delivery service instance.</param>
         public CompanyCommunicatorPretreatFunction(
             NotificationDataRepository notificationDataRepository,
-            NotificationDelivery.NotificationDelivery notificationDelivery)
+            DeliveryPretreatment.DeliveryPretreatment deliveryPretreatment)
         {
             this.notificationDataRepository = notificationDataRepository;
-            this.notificationDelivery = notificationDelivery;
+            this.deliveryPretreatment = deliveryPretreatment;
         }
 
         /// <summary>
         /// Azure Function App triggered by messages from a Service Bus queue
-        /// Used for kicking off the sending notificaiton flow.
+        /// Used for kicking off the notificaiton pretreatment.
         /// </summary>
         /// <param name="myQueueItem">The Service Bus queue item.</param>
         /// <param name="log">The logger.</param>
@@ -52,7 +52,7 @@ namespace Microsoft.Teams.Apps.CompanyCommunicator.Send.Func
                 draftNotificationId);
             if (draftNotificationEntity != null)
             {
-                await this.notificationDelivery.SendAsync(draftNotificationEntity, log);
+                await this.deliveryPretreatment.SendAsync(draftNotificationEntity, log);
             }
         }
     }
